http://jira.jboss.com/jira/browse/JBIDE-2431 Some seam core tests fail with "seam-gen seamgen does not exist" error message.
Some issues are fixed for the tests and the TestCase passes with OK status in my execution environment.
The "seam-gen seamgen does not exist" issue is to be teset on the Hudson or each other build environment (I have no access to these environments)

Also there is a comment in SeamCoreAllTests.java like: "// Seam 2 tests can't run on hudson for now ;(" - but the testCases presented in the issue require Seam 2 to be installed and used for testing.
This comment makes the issue conflicting.


git-svn-id: http://svn.jboss.org/repos/jbosstools/trunk@9134 a97e2381-89e5-4abb-bab3-167db6db766c