JBIDE-15518 Add to CDI model extensions provided without extension service
Added an optional attribute "recognizer" to extension point
org.jboss.tools.cdi.core.cdiextensions that holds qualified name
of implementation of IExtensionRecognizer that determines if
Java project containes CDI extensions that are not supposed to be
registered in META-INF/services/javax.enterprise.inject.spi.Extension.

Tests are added:
1. Check that extension is recognized.
2. Check that bean created by test extension implementation
is available for injection.
3. Check that if dependent project has the same extension as
parent project, exactly one bean is available for injection (this
should be tested for each extension implementation).